Add Mac Definition and Improve Version Definition
This improves the version definition by adding more specific versions.
Actions can now define specific minimum versions. The minimum version
for the Shortcut and the minimum version for the action will be compared
and will warn you about it.

This adds a Mac definition for defining if your Shortcut is Mac only or
not. This is in tandem with actions being able to define if they are Mac
only or not. This will be compared and you will be warned if an action
is Mac only if this is set to false. iOS only actions may be necessary,
not sure.