Fix parser position tracking
This fixes the parser position tracking to improve the accuracy of
parser error or warning messages, as many times they are reported
to be on the wrong line. This was mostly due to jumping extra spaces
to allow for some parsing to be more convenient or cleaner, but it
lead to other functions skipping ahead to far unintentionally.

I am hoping that this hasn't regressed anything or broken anything
that's not showing up in tests, but we'll get to those issues as they
crop up, this is much more logical.