Consolidate how arguments are checked
This moves the logic that checks if we have all the required arguments
and if values provided for an optional argument are the same as the
default value from the type checker to the argument checker.

Why this wasn't already there who knows.

This also improves the warnings and errors for these checks by being
much more descriptive using the information at hand.