Track include line ranges for error reporting
This basically starts creating maps of where included files exist within
the final file that we combine from all of the included files.

When reporting an error, we check if the line is within the range of an
included file and search the matching included files lines for which
line it occurs in the included file.

This allows for very specific error reporting with included files.