Consolidate creating property list values
This was split into three different functions and also usage of the
other functions with no alias to not have a key resulted in usage of the
functions with an empty string for a key when there is an alias function
for that.

This combines all of them so that only plistKeyValue() and plistValue()
are used with variations being based on the data type given.