Implement custom action arguments
This is the initial implementation of custom action arguments.

Custom actions can define what arguments they would like to use and use
them in their body statement.

When a custom action is called, the arguements given for it will be
collected, then the call to the action is replaced with the statement
body with the defined argument names replaced with the given values.

Direct values may sometimes be copied, so not passed by reference.

But variables are passed directly, or by reference, so a custom action
can overwrite the value of a variable if it really wants to.

Also custom actions have no scope limits of course. So if you want to
use a variable you didn't pass you are able to do so.