Implement nil value type
This implements a null or nil value type that can be used in any syntax
where a value is expected.

In the case that "nil" is used when a value is expected, that parameter
will not be filled out (for better or worse!).

This is inspired by the want to clear out arguments that are not
required but don't want to have to provide an empty value of that type
for which some do not exist (dictionary, array, variables/globals).