Handle comments differently
This changes the comments cli argument to be reversed, so instead you
use this argument to include comments in the compiled Shortcut.

This also reevaluates the logic behind this, now we don't collect or
create a token for comments unless the -c or --comments argument is
being used, we just skip over them.