Improve program panic output and include error in debug panics
This should have a link and text encouraging people to make issues, just
never got to this.

This also makes it a lot more explicit what happened exactly, the last
one just looked like a general error that depending on the error could
look like an issue with your code.