Remove matching from tests
I spent a while trying to get this working, but it turns out not only
was this not really working in the first place, but this is kind of a
hassle trying to make sure the files will be an exact match everytime.

Maybe we can do something in the future to make sure Shortcuts is happy
with the compiled result but it's not this.