Remove auto include
I've been thinking about this feature more and I don't think it makes
sense for Cherri. Shortcuts are very procederal, and this kind of thing
only works for an actual programming language. Cherri puts files in the
order they come from the directory, so you'd need to carefully place
files in the dir and at that point, you might as well just use the
`#include` feature.