Add built-in base64File() action
This adds a new built-in action `base64File()`.

This action takes a file path which it uses at compile time to base64
encode the file at the path and insert it into a text block. This is
particularly useful so you can include an audio file in your shortcut
easily and ask playSound() to play an audio file in your project using
this action.