Stop incrementing indexes
It's time to rectify this. We will no longer automatically increment
indexes as this is just not how the Shortcuts runtime works, and so this
will only help when using integer literals, which is not always the case
and will just be confusing.

I don't like it, but we now start counting from 1 in the language from
here on out because of Shortcuts.