Improve action matching algorithm
This only sorts by value matches as that's all that we care about for
ranking.

This adds a competing check additionally, as most of the time we have
matches but they are not in contention. With this in mind, the matches
would then get selected at random, and this seems to be the solution to
that, as they would not have differences in matching values or params,
we should return in this case, as we need to then just use the default
action.