Fix finding delinquent file
This stores the original contents of the file, this way if the error is
on a line that does not exist within the processed version of the file,
we can still find the original line number of the line with the error.

This might not be the best fix for this, as it requires us to keep the
original contents of the file in memory, but this is useful for error
reporting when we need to determine the difference with the processed
content (includes pasted, etc.)

However, we do clear the original contents from memory as with the
contents after we finish parsing, so other stages are not affected.

This also fixes some issues with just simply not detecting the right
included file that the error is on.